Update phpspec and configure travis
Update the version of PHPSpec
Refactor the specs to the new format
Add additional dev dependency for Magento code
Add Mockery as dev dependency even if deprecated
