+++ spec/MageTest/MagentoExtension/EventListener/AfterScenarioListenerSpec.php
@@ -16,38 +16,39 @@
  *
  * @category   MageTest
  * @package    MagentoExtension
- * @subpackage Service
+ * @subpackage EventListener
  *
  * @copyright  Copyright (c) 2012-2013 MageTest team and contributors.
  */
-namespace spec\MageTest\MagentoExtension\Service;
+namespace spec\MageTest\MagentoExtension\EventListener;
 
-use PHPSpec2\ObjectBehavior;
+use PhpSpec\ObjectBehavior;
+use Prophecy\Argument;
+
 
 /**
- * ConfigManager
+ * AfterScenarioListenerSpec
  *
  * @category   MageTest
  * @package    MagentoExtension
- * @subpackage Service
+ * @subpackage EventListener
  *
  * @author     MageTest team (https://github.com/MageTest/BehatMage/contributors)
  */
-class ConfigManager extends ObjectBehavior
+class AfterScenarioListenerSpec extends ObjectBehavior
 {
     /**
-     * @param MageTest\MagentoExtension\Service\Bootstrap         $bootstrap
-     * @param MageTest\MagentoExtension\Service\Config\CoreConfig $coreConfig
+     * @param MageTest\MagentoExtension\Fixture\FixtureFactory $factory
      */
-    function let($bootstrap, $coreConfig)
+    function let($factory)
     {
-        $this->beConstructedWith($bootstrap, $coreConfig);
+        $this->beConstructedWith($factory);
     }
 
-    function it_should_set_core_config($coreConfig)
+    function it_should_clean_away_any_generated_fixtures($factory)
     {
-        $coreConfig->set('test/path/to/config', 'value', 1)->shouldBeCalled();
+        $factory->clean()->shouldBeCalled();
 
-        $this->setCoreConfig('test/path/to/config', 'value', 1);
+        $this->afterScenario();
     }
 }
